automatic_office/requirements.txt
python-docx
xlutils
